Added Abbreviation Functionality
If node is configured to use abbreviations, the user is presented with
more options. User can condense all number related abbreviations into a
single folder (because the number bands that have a number at the
beginning is small). A user can also capitalize all abbreviations in
case an artist insists on lower case spelling of their name.